requirements.txt
langchain
langchain-community
streamlit
faiss-cpu
sentence-transformers
pymupdf
python-dotenv
